+++ fabric/src/main/java/mod/fuji/module/initializer/command_meta/delay/DelayInitializer.java
@@ -1,13 +1,13 @@
 package mod.fuji.module.initializer.command_meta.delay;
 
 import mod.fuji.core.annotation.Unused;
+import mod.fuji.core.command.argument.wrapper.impl.GreedyCommandString;
 import mod.fuji.core.document.annotation.Document;
 import mod.fuji.core.auxiliary.minecraft.CommandHelper;
 import mod.fuji.core.auxiliary.minecraft.ServerHelper;
 import mod.fuji.core.command.annotation.CommandNode;
 import mod.fuji.core.command.annotation.CommandRequirement;
 import mod.fuji.core.command.annotation.CommandSource;
-import mod.fuji.core.command.argument.wrapper.impl.GreedyString;
 import mod.fuji.core.command.executor.CommandExecutor;
 import mod.fuji.core.command.executor.structure.ExtendedCommandSource;
 import mod.fuji.core.event.annotation.EventConsumer;
@@ -46,7 +46,7 @@ public class DelayInitializer extends ModuleInitializer {
     @Document(id = 1751824706971L, value = "Execute a command in seconds.")
     @CommandNode("delay")
     @CommandRequirement(level = 4)
-    private static int $delay(@CommandSource ServerCommandSource source, double time, GreedyString rest) {
+    private static int $delay(@CommandSource ServerCommandSource source, double time, GreedyCommandString rest) {
         String $rest = rest.getValue();
 
         long scheduleTimeMs = (long) (1000 * time);