+++ fabric/src/main/java/mod/fuji/module/initializer/command_meta/for_each/ForEachInitializer.java
@@ -1,12 +1,12 @@
 package mod.fuji.module.initializer.command_meta.for_each;
 
 import mod.fuji.core.auxiliary.minecraft.PlayerHelper;
+import mod.fuji.core.command.argument.wrapper.impl.GreedyCommandString;
 import mod.fuji.core.document.annotation.ColorBox;
 import mod.fuji.core.document.annotation.Document;
 import mod.fuji.core.auxiliary.minecraft.CommandHelper;
 import mod.fuji.core.command.annotation.CommandNode;
 import mod.fuji.core.command.annotation.CommandRequirement;
-import mod.fuji.core.command.argument.wrapper.impl.GreedyString;
 import mod.fuji.core.command.executor.CommandExecutor;
 import mod.fuji.core.command.executor.structure.ExtendedCommandSource;
 import mod.fuji.module.initializer.ModuleInitializer;
@@ -25,7 +25,7 @@ public class ForEachInitializer extends ModuleInitializer {
     @Document(id = 1751823980406L, value = "Execute a command targeted on single-player for each player online.")
     @CommandNode("foreach")
     @CommandRequirement(level = 4)
-    private static int $foreach(GreedyString rest) {
+    private static int $foreach(GreedyCommandString rest) {
         String $rest = rest.getValue();
 
         for (ServerPlayerEntity player : PlayerHelper.Lookup.getOnlinePlayers()) {