+++ fabric/src/main/java/mod/fuji/module/initializer/command_meta/when_online/WhenOnlineInitializer.java
@@ -7,7 +7,7 @@ import mod.fuji.core.auxiliary.minecraft.PlayerHelper;
 import mod.fuji.core.command.annotation.CommandNode;
 import mod.fuji.core.command.annotation.CommandRequirement;
 import mod.fuji.core.command.annotation.CommandSource;
-import mod.fuji.core.command.argument.wrapper.impl.GreedyString;
+import mod.fuji.core.command.argument.wrapper.impl.GreedyCommandString;
 import mod.fuji.core.command.argument.wrapper.impl.OfflinePlayerName;
 import mod.fuji.core.command.executor.CommandExecutor;
 import mod.fuji.core.command.executor.structure.ExtendedCommandSource;
@@ -44,7 +44,7 @@ public class WhenOnlineInitializer extends ModuleInitializer {
     @Document(id = 1755412463665L, value = "Execute the specified command `exactly once`, when the target player `is online`.")
     @CommandNode("when-online")
     @CommandRequirement(level = 4)
-    private static int $whenOnline(@CommandSource ServerCommandSource source, OfflinePlayerName targetPlayer, GreedyString command) {
+    private static int $whenOnline(@CommandSource ServerCommandSource source, OfflinePlayerName targetPlayer, GreedyCommandString command) {
         String $creatorName = source.getName();
         String $targetPlayerName = targetPlayer.getValue();
         String $command = command.getValue();