+++ CHANGELOG.md
@@ -9,3 +9,7 @@
 
 ## Added
 - [command_meta.run] feature: add `command suggestions` for `/run as {player|console|fake-op}` commands.
+- [command_meta.NOT] feature: add `command suggestions` for `/NOT` command.
+- [command_meta.delay] feature: add `command suggestions` for `/delay` command.
+- [command_meta.for_each] feature: add `command suggestions` for `/foreach` command.
+- [command_meta.when_online] feature: add `command suggestions` for `/when-online` command.